==> 2.DataCleaningRevised.ipynb <==
# import the library
import pandas as pd

## 1. Clean the tracklist data

# read the copy file of tracklist and assign it as a dataframe df_tracklist
df_tracklist = pd.read_csv('tracklist copy.csv')

df_tracklist.head() # check the dataframe

df_tracklist.info() # check the data type and null values of each column

# remove extra strings in the url columns
df_tracklist['artist_url'] = df_tracklist['artist_url'].str.strip("{'spotify': ''}")
df_tracklist['track_url'] = df_tracklist['track_url'].str.strip("{'spotify': ''}")


df_tracklist # check the dataframe

df_tracklist.to_csv('tracklist_cleaned.csv',index = False) # save the cleaned dataframe to csv file

## 2. Clean the tracks's features file

# read the copy of the track features file as a dataframe
df_features = pd.read_csv('features copy.csv')

df_features.head() # check the dataframe

df_features.info() # check the data type and the null value of each column. As the data is quite clean, the original dataset will be used later

## 3.Clean the artists_features file

# read the copy of the artists features file as a dataframe
df_artifeatures = pd.read_csv('artists_features copy.csv')

df_artifeatures.head() # check the dataframe

df_artifeatures.info() # check the data type and the null value of each column

# remove redundant strings from the column genres
df_artifeatures.genres = df_artifeatures.genres.str.strip("[]")

df_artifeatures.head() # check the change

# drop the duplicate rows. As there are duplicate artists, there must be duplicated values from the dataframe
df_artifeatures.drop_duplicates(inplace = True)

# check the information of the dataframe there are total 62 rows left
df_artifeatures.info()

# look for empty strings
df_artifeatures.loc[df_artifeatures['genres']== '']

# replace empty strings to 'not known' in the genres
df_artifeatures['genres'].replace('',"'not known'",inplace = True)

# check the change
df_artifeatures.loc[df_artifeatures['genres']== "'not known'"]

# check the null value of the column genres
df_artifeatures['genres'].isnull().sum()

# save the cleaned dataframe to a new csv file
df_artifeatures.to_csv('artists_features_cleaned.csv', index = False)

## 4. Clean the toptracks file

# read the copy of the toptracks file as a dataframe
df_toptracks = pd.read_csv('toptracks copy.csv')

# check the data type and the null value of each column
df_toptracks.info()

# drop duplicated rows. As there are duplicated artists, there must be duplicated top tracks of the duplicated artists
df_toptracks.drop_duplicates(inplace = True)

# check the number of the rows and columns of the dataframe
df_toptracks.shape

# as there are song top tracks having different artist_id with the artist_id which used to extract the top tracks data (show as below). This is because the track is made by two or more artists. This needs to be solved for later schema building.
# there are 3 methods coming up to my mind:
# 1) add those artist_id to the artist_id list and create a new column to distinguish if the artist _id is in my playlist (due to the time contraint, I didn't choose this one)
# 2) change the artist_id into the one which was used to extract the top tracks data, but it would change how the data works from the spotify ( it may involve copyright)
# 3) drop those tracks and remain the top tracks which are created only by the artists within my playlist, I chose this method at the end.
df_toptracks['artist_id'].isin(df_artifeatures['id'])

# remain the top tracks that are only created by the artists within my playlist
df_toptracks = df_toptracks.loc[df_toptracks['artist_id'].isin(df_artifeatures['id'])== True]


df_toptracks.shape # check the number of rows and columns

# save the cleaned dataframe into a new csv file
df_toptracks.to_csv('toptracks_cleaned.csv', index = False)
