==> 3.TableRecreating Revised.ipynb <==
# Recreate the files according to the created schema, utify the column names

# import the library
import pandas as pd

# read all cleaned files as dataframes
df_artist_features = pd.read_csv('artists_features_cleaned.csv')
df_toptracks = pd.read_csv('toptracks_cleaned.csv')
df_tracklist = pd.read_csv('tracklist_cleaned.csv')
df_track_features = pd.read_csv('features copy.csv')


## 1.track_features_fact.csv

# clarify & unify the column names
df_track_features['track_popularity'] = df_tracklist['Popularity']

# drop useless columns
df_track_features.drop(['track_href','analysis_url','type'], axis = 1, inplace = True)

# rename the colum id as a key
df_track_features.rename(columns = {'id':'listtrack_key'}, inplace = True)

df_track_features.columns # check the column names

# reorder the dataframe
df_track_features = df_track_features[['listtrack_key','track_popularity','danceability', 'energy', 'key', 'loudness', 'mode', 'speechiness',
                                       'acousticness', 'instrumentalness', 'liveness', 'valence', 'tempo',
                                        'duration_ms']]

df_track_features.head() # check the changes

# save the revised dataframe as a new csv file
df_track_features.to_csv('listtracks_features_fact.csv', index = False, encoding = 'utf-8')

## 2.toptracks_popularity_fact

# choose the columns for the fact table
df_toptracks_new = df_toptracks[['track_id','track_popularity']]

# rename the column names
df_toptracks_new.rename(columns = {'track_id':'toptrack_key'}, inplace = True)

df_toptracks_new.head() # Check the new dataframe

# save the dataframe as the troptracks' popularity fact file
df_toptracks_new.to_csv('toptracks_popularity_fact.csv', index = False,encoding = 'utf-8')

## 3.artists_features_fact

# look for the data I need for the fact table
df_artist_features.head()

# extract the columns needed for the fact table and assign into a new dataframe
df_artist_features_fact = df_artist_features[['id','followers','artist_popularity']]

#rename the column name
df_artist_features_fact.rename(columns = {'id':'artist_key'}, inplace = True)

#save the dataframe as a csv file of the fact table of artists' popularity
df_artist_features_fact.to_csv('artists_features_fact.csv', index = False, encoding = 'utf-8')

## 4.my_artists_DIM

df_artist_features.head()# look for the data needed

# extract the data for the artist dimension table
df_artist_dim = df_artist_features[['id','artist_url','artist_name','genres']]

# rename the id column as key
df_artist_dim.rename(columns = {'id':'artist_key'},inplace = True)

# save the dataframe to the dimension table as a csv file
df_artist_dim.to_csv('my_artists_DIM.csv', index = False, encoding = 'utf-8')

## 5.toptracks_DIM

df_toptracks.head() # look for the data needed

# extract the data needed for the toptracks dimension table
df_toptracks_dim = df_toptracks[['track_id','track_url','track_name','artist_id']]

#rename the column names
df_toptracks_dim.rename(columns = {'track_id':'toptrack_key','artist_id':'artist_key'}, inplace = True)

# save the new dataframe as a csv file for the toptracks dimension table
df_toptracks_dim.to_csv('toptracks_DIM.csv', index = False, encoding = 'utf-8')

## 6.playlist_tracks_DIM

df_tracklist.head() # look for the data needed

# unify the column names as lower cases
df_tracklist.columns = df_tracklist.columns.str.lower()

# extract the data needed for tracklist dimension table
df_tracklist_dim = df_tracklist[['track_id','track_url','track_names','artist_id']]

# rename the column names, unify the names
df_tracklist_dim.rename(columns = {'track_id':'listtrack_key','track_names':'track_name','artist_id':'artist_key'}, inplace = True)

# save the dataframe as a csv file for the track list dimension table
df_tracklist_dim.to_csv('listtracks_DIM.csv', index = False, encoding = 'utf-8')
